выполнить 8-й шаг

+++ src/main/java/hexlet/code/service/TaskStatusService.java
@@ -1,11 +1,13 @@
 package hexlet.code.service;
 
-import hexlet.code.dto.TaskStatusCreateDTO;
-import hexlet.code.dto.TaskStatusDTO;
-import hexlet.code.dto.TaskStatusUpdateDTO;
+import hexlet.code.dto.task_status.TaskStatusCreateDTO;
+import hexlet.code.dto.task_status.TaskStatusDTO;
+import hexlet.code.dto.task_status.TaskStatusUpdateDTO;
+import hexlet.code.exception.MethodNotAllowedException;
 import hexlet.code.exception.ResourceNotFountException;
 import hexlet.code.mapper.TaskStatusMapper;
 import hexlet.code.model.TaskStatus;
+import hexlet.code.repository.TaskRepository;
 import hexlet.code.repository.TaskStatusRepository;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
@@ -17,6 +19,9 @@ public class TaskStatusService {
 
     @Autowired
     private TaskStatusRepository statusRepository;
+
+    @Autowired
+    private TaskRepository taskRepository;
 
     @Autowired
     private TaskStatusMapper statusMapper;
@@ -49,6 +54,10 @@ public class TaskStatusService {
     }
 
     public void delete(Long id) {
+        if (!taskRepository.findByTaskStatusId(id).isEmpty()) {
+            throw new MethodNotAllowedException("You cannot delete a status. The status is associated with tasks.");
+        }
+
         statusRepository.deleteById(id);
     }
